Cache static part of ReflectionParamInfo
Summary:
Somehow `findClosestGetterOnAbstractClassOrInterface` shows up in profiler.

So cache lookup then.

On my laptop `QueryCommandIntegrationTest`:
Before this diff: 42s
With this diff: 35s

This diff only speeds up tests, but not production Buck.

The essential part of this diff is addition of:

```
private static final ConcurrentMap<Method, StaticInfo> staticInfoCache =
      new MapMaker().weakValues().makeMap();
```

The rest is fix of compilation errors.

Reviewed By: bobyangyf

shipit-source-id: 7396e3e971