Allow apps to depend on other apps in iOS UI tests (#2346)
Summary:
To test correct inter-op with third party applications, we often us helper "sidecar" apps in our UI tests at Dropbox. During the UI tests, they need to be built and installed together with the main app. This works well from command line, but currently fails in Xcode.

This PR enables Buck to build these additional apps and include them as a part of the XCUITest bundle both in Xcode, and in the CLI.
Pull Request resolved: https://github.com/facebook/buck/pull/2346

Reviewed By: williamtwilson, swgillespie

Pulled By: swgillespie

shipit-source-id: 74e6028be8