Do not use backslashes on non-Windows
Summary:
Test added in D13370072. AFAIU supposed to test than on Windows it uses forward slashes.

But on non-Windows backslash is not a separator char. And we are not supposed to use backslash as a file name.

Following diff adds an implicit assertion that filename (or path) does not contain backslash on non-Windows, and that assertion fails without this diff.

Reviewed By: gabrielrussoc

shipit-source-id: fb520b91f0