Abstract out AbstractSkylarkFileParser and utilize to create SkylarkPackageFileParser
Summary:
The majority of the existing `SkylarkProjectBuildFileParser` class can be reused for parsing PACKAGE files in Skylark.

To make this work, we extract out the core functionality of `SkylarkProjectBuildFileParser` to `AbstractSkylarkFileParser` and then extend it to form the concrete classes
- `SkylarkProjectBuildFileParser` (existing)
- `SkylarkPackageFileParser` (new)

The `SkylarkPackageFileParser` returns a `PackageFileManifest` which contains the `PackageMetadata` read from the PACKAGE file, as well as any files included, configs, etc.

Reviewed By: stepancheg

shipit-source-id: 4360ca3ddc