Make BuildFileParsePipeline generic
Summary:
In addition to a parse pipeline for Build files, we also need a parse pipeline for PACKAGE files. The functionality is identical except for the parser used and the object type parsed.

Based on all the previous work to make file parsing generic, we can now make the file parse pipeline generic.

`FileParsePipeline` performs all the functions previously performed by `BuildFileParsePipeline`, which now is just a concrete casted implementation.

Reviewed By: stepancheg

shipit-source-id: 05b1e4993f