Add --show-outputs flag
Summary: Throws exception if a rule supports multiple outputs but the flag wasn't passed. However, if a rule supports multiple outputs, the flag wasn't passed, but the default output label was passed, that means the user is using a multi-output compatible rule but not using the multi-output feature, so don't throw in this case. Additionally, I think `--show-outputs` should behave the same as `--show-output` for non multi-output compatible rules (i.e. print the one output), so if `--show-outputs` is passed for a rule that doesn't support multiple outputs, it doesn't throw an exception and instead prints the one output path.

Reviewed By: philipjameson

shipit-source-id: 9410a831f1