Use Sorted collections of Artifacts in Actions for stable ordering
Summary:
We iterate over the Artifacts of the Actions when computing rulekeys.

Some of this then reflectively enters BuildRules and such in a complicated manner.
We should make the order of these Artifacts stable so that the iteration of it when calculating rulekeys are stable, just like how we add Build Deps for BuildRules to ensure consistent hashing.

Reviewed By: philipjameson

shipit-source-id: 747a85e011