Make shared lib return type optional
Summary:
Values returned from CxxBuckConfig should be Optional, so that they can be correctly chained. The need arises for flavored platforms (e.g., "cxx" and "cxx#macosx-x86_64"). When composing values for the platforms, this means we need to be able to distinguish between no option set, in which case we would pick the value from the non-flavored platform. That's why the return types have to be Optional.

NOTE: This diff does not change behavior.

Reviewed By: selat

shipit-source-id: bf58a92135