Add ULEB128 utility
Summary:
We need the ability to convert from/to ULEB128 encoding which we add here. Note that we do not want to re-use the pre-existing LEB128.java code which is part of `third-party/.../android/dex` for two reasons: 1) It uses a custom ByteInput interface 2) it uses 32bit rather than 64bit values.

For better compatibility with ld64, we implement the utility here.

Reviewed By: stepancheg

shipit-source-id: b497ef5d19