Reader for LC_SYMTAB command
Summary: Implements a reader for the LC_SYMTAB command. We need to modify the symbol table to zero out all symbol addresses for the dylib stub (so that the dylib stub has the same contents as long as the ABI stays the same, even if the implementations of the symbols change).

Reviewed By: selat

shipit-source-id: eeba1224d5