Add dylib stub scrubber
Summary:
The role of the dylib stub scrubber is to remove all non-deterministic parts of the dylib stub. We want to ensure that the dylib stub generated from a dylib stays exactly the same, even if the implementations of the symbols change.

For the purposes of linking, users of a dylib only need to care about the exposed ABI, not the internal implementations. Dylib stubs allow us to not recompile the full tree of dylibs, if the underlying ABIs don't change. This is useful to speeding up incremental builds.

Reviewed By: selat

shipit-source-id: 76ac9c0260