Add support for deps for @ExcludeFromRuleKey
Summary:
The ExcludeFromRuleKey annotation already provides the ability to opt into serialization and inputs behavior but it does not provide the ability to derive deps in the same way.

In this diff, we add that support, in the same way it works for inputs. This ability is useful when rules compute ABIs where they depend  on the source inputs (i.e., depend on the source `BuildRule`s) but the rulekey should not include the source input, by definition (as the ABI is a derivative).

NOTE: There are no behavioral changes, as no changes are needed for pre-existing tests. This is a purely additive change.

Reviewed By: selat

shipit-source-id: 22f8d8b539