Add rule to strip dylib stub
Summary:
This adds a build rule which scrubs a dylib stub to remove symbol addresses in the symbol table + export tables and reset the UUID. As a consequence, if the source dylib only changes the implementation of its symbols, the resulting dylib stub will not change.

This is a two step process: firstly, we use the "strip" command and then scrub any non-deterministic parts (i.e., addresses + UUID).

Reviewed By: selat

shipit-source-id: 4ba90510ef