Add support for shared library interface for Apple platforms
Summary:
Add support for shared library interfaces for Mach-O libraries (i.e., dylibs). This diff means that if the support is turned on via "cxx.shlib_interfaces = enabled", then dylib stubs will be used for linking.

The benefits here are that incremental linking time would be eliminated completely, if the source dylib's ABI does not change (i.e., only implementations change). This is the first iteration of the feature as I'm planning to add support for tbd (text based stubs) later on.

We had to inject the "strip" tool as there's an interesting dependency: the shared lib params needs to get values out of the `CxxPlatform` but the `CxxPlatform` itself contains the shared lib params.

NOTE: This does not change existing behavior unless explicitly opted in via "cxx.shlib_interfaces = enabled".

Reviewed By: selat

shipit-source-id: 77ef651da6