Fix Mach-O export trie reading
Summary:
This fixes an oversight introduced in the initial code. We were accounting for the size of the export info twice which lead to broken reading of properly nested tries.

In the code we have:

```
if (exportInfoSize > 0) {
  exportInfo = Optional.of(readExportInfo(byteBuffer));
}
```

The code above would already advance the buffer position, so there's no need to add `exportInfoSize` again, thus causing an incorrect offset.

The reason why the bug was not caught by the original diff is because the tree only had two levels.

Reviewed By: mattjgalloway

shipit-source-id: 9c6f6c6a3e