Fix the flavor propagation point for bundles
Summary:
In the previous iteration, we just discover whether a flavor is missing and add it in.
While if we were just dealing with bundle rules, this would be fine, but the asset rules and others
will also get flavored in the same way causing an IllegalStateException within the action graph rule
resolver.

The fix is to manually check and resolve the correct flavor later on when building the rule by verifying
that the `default_platform` does supercede the buck-configured cxx.default_platform.

Why not use the existing `ApplePlatforms::getCxxPlatformForBuildTarget`? Because that also takes into
account any flavor already on the passed in `BuildTarget`. Instead, we want to skip that first step.

Reviewed By: stepancheg

shipit-source-id: e67983200f