Add outs parameter to genrule
Summary:
Adds `outs` parameter to genrule as a map. Populates the field in `Genrule` but doesn't use it yet.

`outs` is mutually exclusive with `out`. `out`'s user-visible behavior shouldn't change as a result of this diff. Instead of integrating `out` logic into `outs`, chose to implement the two separately in parallel so the `out` logic can be easier to remove once things have been migrated to `outs`.

Reviewed By: philipjameson

shipit-source-id: c13c99a7a7