Allow target_sdk_version on binaries for Obj-C, test it for tests
Summary:
I previously added support for `target_sdk_version` in `apple_library`. However, we should also support it in `apple_binary` and `apple_test`. This diff adds support for `apple_binary`, for Objective-C only. It turns out that `apple_test` already works, as the `apple_library` code is reused, so I added a test case ensuring that continues to be the case. Then, I updated the docs to include `target_sdk_version` on all three pages.

`apple_binary` is a little tricker for Swift, because it's using a different code path to compile Swift code. I will add that support in a separate diff, to keep things clear.

shipit-source-id: 90505524d5