Support target_sdk_version for Swift in apple_binary
Summary:
The `apple_binary` Swift compilation goes through a different code path currently, so to support `target_sdk_version`, we need to do it a little differently.

The "Swift library" code doesn't appear to be coupled to Apple platforms, so avoiding any direct references to "Apple..." classes, and adding a similar `getTargetSdkVersion()` method to the Swift args. Then, we can set that value in the Apple code.

shipit-source-id: bfd3453a3e