Fix --out param for buck targets that output a directory
Summary: Currently, if you use the --out param when building a buck target that outputs a directory, buck will create an empty directory with the same name at or in the path passed to --out, but doesn't copy any of the source directory's children. This diff changes the behavior to copy the whole directory contents.

Reviewed By: philipjameson

shipit-source-id: 9a17a69c38