Ensure RE works correctly for isolated buck invocations.
Summary:
This diff attempts to enable RE to correctly execute "isolated Buck" commands
(i.e. commands that make use of the `--isolation_prefix` command line option).

The key part that appears to be missing is passing over the "base buck out"
directory, that is calculated by the Python execution wrappers around Buck and
(in local execution) passed to `buckd`. This change collects up the value of
that system property and pipes it through to the remote execution via the
`trampoline.sh` script.

Took the opportunity to simplify/tidy-up the original isolated buck path-rewriting
stuff and also extend it to cover embedded cell cases.

Reviewed By: cjhopman

shipit-source-id: 19872f7913