Libraries should be modular regardless of whether they are focused
Summary:
Problem: If you try to run a Swift class that imports an unfocused library, you will get a "library not found" error. The reason for that is because we do not create sym links for header files if the library is unfocused.

Fix: Remove `isFocusedOnTarget` check so that header sym links are created regardless of whether the libraries are focused.

Reviewed By: williamtwilson

shipit-source-id: d62796a871