Support binary resources mode in Robolectric tests
Summary:
Augments `RobolectricTestDescription` to also run `aapt2` iff `use_binary_resources` is specified on the corresponding target.

If binary resources are enabled, we omit the output of `DummyRDotJava` and instead use the output from `aapt2` and include a unit test options file detailing where the output resource apk is.

Fixes #1549

Reviewed By: asp2insp

shipit-source-id: 93f6a28eb97c416770090c15370df53180046aa9