Support aapt2 locale filtering.
Summary: Adds an option `aapt2_locale_filtering` to `android_binary`. When set to `True` and one or more locale is present in the `locales` attribute, Buck will filter locales by passing the `-c` option to `aapt2 link` instead of filtering by copying resources into a filtered resource tree. For apps with large numbers of resources, this can be significantly faster.

Reviewed By: asp2insp

shipit-source-id: 48fef670ccb4cc823d72cd47d109236461e89689