python: remove unused path root from prebuilt module dirs
Summary:
Internally, Buck maintained prebuilt python libraries via a multimap, mapping a path
root to the prebuilt contents to be placed that location.  However, this was never exposed
to users and path roots were always set to the empty path, meaning they were effectively
unused.

This diff removes the path root, replacing the multimap with a set of prebuilt libraries, which
allows simplifying use of prebuilt libraries, but also simplifies `PythonSymlinkTree` and
`SymlinkTreeMergeStep`.  Longer term, this also helps with some refactoring to start
compiling bytecode in Buck.

shipit-source-id: de5b22efa6862d6dc8f97c03f1df8b512892412b