python: minor refactoring of `PythonUtil.getAllComponents()`
Summary:
We use `PythonUtil.getAllComponents()` to traverse a graph from a Python
binary/test to collect all the components that we need to package with
it.  Previously, we'd special case the root node by explicitly passing
in it's package info, then start the traversals from the deps.  This diff
refactors this a bit so that we wrap the top-level root nodes in a
`PythonPackagable` to allow starting the traversal at the root and avoiding
needing to pass in additional info to special case the root node.

This is a noop as of now, but will make some future changes to Python rules
(e.g. support compiling Python sources in Buck) easier.

shipit-source-id: ed3691fc57a770edb51caf2fc848a0bc0b408be8