python: add `PythonComponents.forEachInput()`
Summary:
Add a `PythonComponents.forEachInput` method to abstract accessing
`SourcePath`s wrapped by `PythonComponents`, whether they come from
a directory of sources or a static map.  This will eventually be used to replace
the separate casing in in-place and packaged Python binary implementations
for the separate directory-of-source and static-map cases.

shipit-source-id: ec18af1e5b45ac10acc1d4c264245d9402bdc768