Add support for location macro with multiple outputs
Summary:
This already mostly worked due to `HasSupplementaryOutputs`. They both use the square bracket syntax.

We can't really tell from the square bracket syntax whether a target is a supplementary output or target with output label, so use `instanceof` to check. This means that `HasSupplementaryOutputs` needs to be mutually exclusive with `HasOutputLabel` (which was implied before, but this makes it more "official")

Can follow up with some renaming in `BaseLocationMacro` so it doesn't refer to supplementary outputs directly and instead of refers to the square bracket syntax

Reviewed By: philipjameson

shipit-source-id: 884e8571ee2f280b7c01836e93ce0950c5bee7be