SymlinkTree: add a `MappedSymlinkTree` specialization
Summary:
A future change updates `SymlinkTree` to accept a more general way of specifying
the symlinks it should create, rather than a static file map.  However, a number of cases
rely on the `SymlinkTree.getLinks()` helper to access the static file map they used to
construct it.

To make the subsequent change to `SymlinkTree` easier, this adds a `MappedSymlinkTree`
specialization which preserves the existing behavior and updates cases that rely on it
to use this new class.

shipit-source-id: 112d56af33010563937f142c3f4df3553b008c33