Exit remote execution when job was stolen by local machine
Summary:
When a job is stolen by a local machine using HybridLocalStrategy, remote execution of that job will now stop as soon as it can, avoiding unnecessary work in COMPUTING_ACTION, UPLOADING_INPUTS, and UPLOADING_ACTION. We check at state changes to see if the job has been cancelled, and throw an exception which stops the execution of the async transforms.
I also changed the code to do the same during the EXECUTING step.

Reviewed By: rajyengi

shipit-source-id: 25ae2c475edfc8d242bbbf747e0d3be05cc9464e