Use the DaemonicParserState in ParsePipelineTest
Summary:
`ParsePipelineTest` is an integration test of the parsing flow that includes parsing files, creating nodes, caching, etc. It utilizes a `TypedParsePipelineCache` for caching of build manifests, nodes, etc. This makes it impossible to to test invalidation and any re-parsing that needs to be performed.

To fix this, we use the real `DaemonicParserState` which contains real implementations of caches.

However, this fails 3 existing tests, which relied on a fake cache. Specifically,
- `exceptionOnMalformedRawNode`
- `exceptionOnSwappedRawNodesInGetAllTargetNodes`
- `exceptionOnSwappedRawNodesInGetTargetNode`

all error in failing to create a valid build target via `UnflavoredBuildTargetFactory` due to invalid node content or a "swapped" build file path.

`DaemonicParserState` creates build targets during the caching process, so the tests fail earlier than previous intended, making the `expectedExceptions` incorrect. We could just expect an exception earlier, but that defeats the purpose of the test.

Realistically, these tests verified the logic in `UnflavoredBuildTargetFactory`, so I wrote new tests for UnflavoredBuildTargetFactory and removed them from `ParsePipelineTest`.

Reviewed By: bobyangyf

shipit-source-id: d02fcb68b4edaf34810bb7f81be993145ef574be