Respect `default_platform` attribute for dependencies of apple_bundle
Summary: Right now we're assuming that we can just pick up the default platform from the buckconfig, but we want to be able to grab that information from a per-target attribute. For repos without `cxx.default_platform`, this fixes a bug where we would build applications for `iphonesimulator-x86_64` but we would build extensions for the default platform.

Reviewed By: cjhopman

shipit-source-id: 2f71cfedc1e4c5d4fbe8aec98eb3e3aeb4d99d05