Enable opt-in for genrule remote execution
Summary:
This commit adds a new property to the 'genrule' rule, 'remote', which opts-in to remote execution for that genrule. If a genrule sets this property to True and the genrule itself is serializable, Buck will execute that genrule remotely just like every other modern build rule.

To ease the migration process, this commit does a few conservative things:
   1) The 'remote' flag is False by default, ensuring that no genrules magically start remote executing (since many of them don't work on RE)
   2) Even if 'remote' is set to True, Buck will ignore it unless the config key 'remoteexecution.use_remote_execution_for_genrule_if_requested' is present and true in the user's buckconfig. This config is also false by default, so even if we do add 'remote = True' rules, they won't run on RE unless enabled.

Reviewed By: irenewchen

shipit-source-id: d71a5291716f45b9b233be8eb7ade17dc97820e5