Don't crash when minSdkVersion is not numeric.
Summary: If an AndroidManifest.xml file contains a non-numeric value in minSdk, buck project would crash with a NumberFormatException. Some toolchains include non-numeric values in this attribute because of additional processing. Ideally, we'd understand the result of this processing, but as a stopgap, just avoid considering non-numeric values when determining the minimum.

Reviewed By: IanChilds

shipit-source-id: 8f784ebeaeb45434cf062def75fc17f28b5933eb