mv *BuildTarget and *UnconfiguredBuildTarget to upper package
Summary:
Move classes from `model.impl` package to `model`.

There's no profit in separating interface and implementation for these types, and having them in separate packages creates a lot of headache with dependency cycles.

This is part of the work of simplyfing `*BuildTarget` classes structure, we need to:
* merge `UnconfiguredBuildTargetView` and `UnconfiguredBuildTarget`
* remove interface/inheritance between various `*BuildTarget` classes where it is not needed

This diff only moves types, nothing more.

Reviewed By: bobyangyf

shipit-source-id: 7debfbf462c95bfdf58f2a20598e22d506d91bbc