When querying, add computed attributes to the dictionary that is output
Summary:
For a good UI, when we query a node, we want to return attributes computed using the package, not just the raw attributes defined in the BuildFileManifest. Otherwise this would lead to confusion and not provide a clear reason of all values on a target.

Specifically with regards to visibility/within_view.

Example:

.buckconfig
```
[parser]
  enable_package_files = true
  enable_inheriting_package_files = true
```

/PACKAGE:
```
package(
    visibility = [
        "//foo/...",
    ],
)
```

/BUCK:
```
java_library(
    name = "bar",
    ...
)
```

The following command:
```
buck query //bar --output-attributes name visibility
```

Previously:
```
{
  ":bar" : {
    "name" : "IGNetworking",
    "visibility" : []
  }
}
```

Now:
```
{
  ":bar" : {
    "name" : "IGNetworking",
    "visibility" : [  "//foo/...", ]
  }
}
```

Reviewed By: stepancheg

shipit-source-id: 307878d0636c66af44ba3cf38231d7da3b992b67