Don't ignore files when tidying cache artifact directories
Summary:
When decompressing a cache artifact, we tidy up the destination directory by clearing out any file that isn't present in the archive. Previously, this code used `getDirectoryContents` to enumerate the contents of the destination, which skips files that are ignored by Buck, which in turn would cause Buck to skip the deletion of certain files (e.g. .pyc) that it needs to delete for correctness.

This commit uses `getFilesUnderPath` on `ProjectFilesystemView` to accomplish the same thing, except that `getFilesUnderPath` includes paths that are otherwise ignored by Buck.

Reviewed By: simpkins

shipit-source-id: 099362eeebf408bec8960cd37c86c79b8e3c5454