Store the `definingPath` on a `VisibilityPattern`
Summary:
With `packages()` and inheriting `PACKAGE` files, targets can have `VisibilityPatterns` that are not defined in the target's build file. Store the `definingPath` along with all ``VisibilityPatterns` so that we have information on where the pattern was derived from.

This is used in the next diff to output more structured query data, so that output data is more user friendly/helpful.

Reviewed By: bobyangyf

shipit-source-id: 57cbb4c2797071ee28977711fd6c56f89c0f2b95