Build all modular nodes to generate swift modulemap
Summary:
Swift module maps don't currently work in projectv2 because of a change we made into how projects were generated.

In the current buck project, all targets in the targetGraph are looked at and can have stuff performed on them, even if they're not "focused" on; as a result, targets not focused on would have their generated module map files generated in `buck-out` basically as a side effect of this.

In projectv2, we changed this behavior because it was redundant and not necessary -- we do the focus up top before generating anything Xcode-related, so as not to do any work we don't need to (if you're not `--focus`'d on it, why care about it?). This works pretty well, but then broke the above-mentioned side effect in that Swift modulemaps were not generated so click through works ... kind of, sometimes (despite the fact we do all the necessary header generation for objc).

There is also a difference between how module maps are created in buck project vs. buck build -- buck build put them in `buck-out/gen/{cell_relative_path}` while project put them in `buck-out/gen/_p/...`. This is unnecessary because we are already putting them in one place, so it seems redundant not to just piggy-back off this functionality for free.

This diff resolves both above problems for projectv2.

- After generating a project, we iterate through all of the target graph nodes, get all of them that are modules, and created the specific build flavor we need of the target to generate the symlinks.
- We pass these build targets to the buck build phase run at the very end of project generation (which happens already for targets like genrule files) which will symlink everything for us for free.

As a result, we now see in `buck-out` all of the generated module information.

For the build flavor, I defaulted to `iphonesimulatorx86_64` instead of creating both by including `arm` because headers should be the same for both, so I didn't think it was worht caring about this more than we need to.

The follow-up diffs to this will clean up the existing logic and change the `SWIFT_INCLUDE_PATHS` to point to the buck build directory.

shipit-source-id: 15666c0dd2b945efeef79e06581b443b43e74e71