Add additional constructor method to allow modifying skylark values
Summary:
Sometimes it's useful to modify a skylark value, validate it, etc before passing it into the provider.
Allow a static method to be specified that returns an instance of `BuiltInProviderInfo` and takes arguments
much like the constructor normally does. If specified, this method will be called rather than the constructor
in skylark

Reviewed By: bobyangyf

shipit-source-id: 7c51e576c70435388c1f2223938f2f7708315566