Support printing multiple outputs in BuildReport
Summary:
For rules with multiple outputs, this prints something like:

  "success" : true,
  "results" : {
    "//:rule_with_multiple_outputs" : {
      "success" : true,
      "type" : "BUILT_LOCALLY",
      "outputs" : {
        "output2" : [ "buck-out/gen/rule_with_multiple_outputs__/out2.txt" ],
        "output1" : [ "buck-out/gen/rule_with_multiple_outputs__/out1.txt" ],
        "DEFAULT" : [ "buck-out/gen/rule_with_multiple_outputs__/out1.txt", "buck-out/gen/rule_with_multiple_outputs__/out2.txt" ]
      }
    }
  },
  "failures" : { }

For JSON output, the code would be simpler if we can directly change `output` to behave the same way as `outputs` in the existing code. However, there are projects that are parsing `output`, so unifying the behavior would break these projects.

For default outputs, added some special logic to print DEFAULT to be consistent with select. Also open to changing `OutputLabel#toString` to print DEFAULT instead of <default> so no special logic would be required. I'll add a preconditions check so that users can't make a label with the default string.

Reviewed By: bobyangyf

shipit-source-id: fb1331e577a2ccd46bbc26137656a0ae689ebe1b