python: add `__init__.py` modules for `.pyi` extensions too
Summary:
The change to coalescing python components (e.g. modules) at build
time changed how `__init__.py` files were automatically added by Buck
when missing, and no longer added when a package didn't contain any
Python modules (i.e. `.so` and `.py`).

It turns out that some type-checking tests in the wild expect these for `.pyi`
as well and, while it's not clear if this is the best way to handle this case,
restore this behavior for now.

Reviewed By: andrewjcg

shipit-source-id: 500e86a9a07fd61558230a517341c9c46f81ffe2