Store the last counter value when generating Global IDs
Summary:
The biggest bottleneck for project generation on projectv2 has been writing the project files. After playing with this for a day or so, I've discovered the issue to be the generation of Global IDs.

Global IDs are generated based on two components of a `PBXObject` -- the `stablehash` which is usually a constant value `0` for all classes that are not files, which use the hashcode from their file name. In order to handle clashes, it also uses a counter and a while loop which it runs through until it finally finds a GID that is not used. This worked fine for projectv1 because each buck package was its own pbxproject and used its own GID generator, so the clashes were minimal -- the biggest clashes came from Targets and build phases, of which there would only be a handful of.

In projectv2, with just the one project file, we are hitting a scaling problem. For a large project, we can have a ton of different conflicting build phases and targets, which means every time we generate a GID, we have to loop through every previously used counter to find a non-clashing one. This is prohibitively expensive.

To solve this, I'm introducing a hashmap here that will hold the next available counter object  each object type so we can quickly find the next counter object and reduce the amount of loops to 0. The key for this hashmap is the PBXObject.isa() string and is only considered if the `hash` is not 0 -- which will be for all objects that are not file references, for which this won't matter because the the file name is used for the hash and should only ever have one loop anyway.

I'm still keeping the hashset around of conflicting IDs just in case this isn't foolproof, but theoretically we shouldn't need this anymore.

Reviewed By: williamtwilson

shipit-source-id: 93b652cf36a8cd02cb736122d423ada44f05a51b