Allow args to be populated from action.args()
Summary: Currently the Bazel API requires doing `ctx.actions.args()`, then adding to that. Many times we just want to populate `CommandLineArgs` objects directly when creating the object with args(). Add a single value, and iterable interface to args() to let us make it easier for users to instantiate things.

Reviewed By: bobyangyf

shipit-source-id: dbe68a78f618212763a076666a8a0cb2ef4bba8b