Implement AddToRuleKey for CommandLineArgs
Summary:
Implement `AddToRuleKey` for `CommandLineArgs`. This gets a little hinky with `Artifacts` (namely that we don't want to assume a file that hasn't been written yet exists on disk). Artifacts that are in these CommandLineArgs should have already been declared as inputs/outputs of other actions, so we /should/ correctly take file changes into account when necessary. (Note that CommandLineArgs are completely detached from Actions, so we can't/shouldn't introduce an 'inputs' vs 'outputs' concept)

See the comment in the code below for how we get around this for now.

Reviewed By: bobyangyf

shipit-source-id: 35a2f20c273fb7648a6cc3a9d8a2631b1c8826e0