Add integration test to surface bundle flavor issue
Summary:
We've run into a problem with removing `cxx.default_platform` where we get an error about how two different bundles share the same binary. The problem is that they aren't two different bundles - they are the same bundle, but one has a flavor and one does not.

This diff is meant to surface this error in an integration test (which is currently disabled) that we can then work on fixing.

Reviewed By: miaoyipu

shipit-source-id: e59664bd0daf8e6ae79c7a3acd13dc9e427cd0cd