Introduce ProviderBackedUnresolvedAppleCxxPlatform
Summary:
Finally use `apple_toolchain` and `apple_platform` rules.

When `apple.toolchain_target` is set in the config, the value should be a build target which will be used to define `AppleCxxPlatform` instead of current logic to search and configure the platform.

It provides much more flexibility than current version. For example, the whole toolchain can be expressed in native Buck targets which can download and setup the toolchains from some external storage.

Reviewed By: cjhopman

shipit-source-id: c70b9ad223a91886daf4de72eeda7b86c94eee42