Make trim uber r dot java depend on dex groups instead of predexed libraries, again
Summary:
This had a nasty bug the first time it landed, and was reverted.

The only change in the implementation is this:
    // Make sure this index is larger than the last index used for secondary dexes in the root
    // APK module, so that an existing secondary dex can't be overwritten in the merge step.
    int finalDexGroupIndex = dexGroupMap.get(apkModuleGraph.getRootAPKModule()).size() + 1;

Which replaces `dexGroupMap.size() + 1` as the secondary dex index used by the new dex group used for R.java classes.

Because the wrong index was used, in cases where `allow_r_dot_java_in_secondary_dex = True`, this dex group could reuse an existing secondary dex name. For example, if there is only a single APK module, it would overwrite `secondary_2_1.dex.jar`.

Reviewed By: IanChilds

shipit-source-id: cdec2e1288fb82eafcf8a8dd0902dec3fd6dca38