Add OutputArtifact and as_output() to Artifact
Summary:
In some cases we want to be able to infer inputs/outputs for an action. We cannot just use bound/unbound for this, as we might end up accidentally declaring a file as an output, when the user just forgot to bind that artifact before using it in the inferring code.

Add an `as_output` function to handle this, and return a thin wrapper that will indicate that this usage of the artifact is as an output. Bound artifacts will always be able to be assumed to be an input

Reviewed By: bobyangyf

shipit-source-id: 4ffaf5aeff0bb960eae51b9012d18e916460eb5f