Require that arguments to CommandLineArgs are either bound artifacts, or output artifacts
Summary: We want to be able to infer inputs/outputs from arguments provided by users. To do this, we need to be able to differentiate between artifact types. Assume that bound artifacts are inputs. Unbound artifacts need to be explicitly marked as outputs with as_output, and we need to be able to store that in commandlineargs. We do this because otherwise a user who just forgot to bind and artifact to an action before calling run() might inadvertently declare it as an output to run(), rather than as an input (with it being an output of the missed action)

Reviewed By: bobyangyf

shipit-source-id: 351e5286422e28ee0d007a026ef95b9637e201ba