Export BuiltInProviders from plugins
Summary: Some plugins, e.g. dotnet, have providers that need to be made available to users in the parser. Unfortunately this generally crosses over plugin boundaries, so we cannot make the parser directly depend on the various features. Instead, add a new interface that exports various built in providers that should be available to users. This will be used for C# libraries, and C# toolchains initially.

Reviewed By: bobyangyf

shipit-source-id: c7038bc22f2c91b78fc8127bd31603e9336f628a