Add LegacyToolchain rule
Summary: Add a new rule, 'legacy_toolchain'. This rule is compatible with RAG, and proxies providers from an underlying Toolchain. Names come from the various Toolchains. With this, users can specify a legacy_toolchain rule, and depend on it just like any other target. Aside from this one rule/description, we shouldn't have to have any other special cased toolchain logic propagate into UDR or most of the RAG. This will likely change as we solidify our concept of 'toolchains' in the future.

Reviewed By: bobyangyf

shipit-source-id: 89d13b5f6ed1d29d92e393bcd2077dc16831e1cb