Add support for getting output name by output label in genrule
Summary:
Add `getOutputName(OutputLabel)` to `HasOutputName`

This diff adds an implementation for `getOutputName(OutputLabel)` for genrule and calls this method in `DefaultSourcePathResolver`

Ideally we'd want to remove `getOutputName()` and have `getOutputName(OutputLabel)` only

To do that we need to pipe the output label from `AbstractResourcesParameter`

and `DefaultSourcePathResolver` (see comment below)

We can do that if we elevate `getTargetWithOutputs()` from `DefaultBuildTargetSourcePath` to `BuildTargetSourcePath`. Since `OutputLabel` is now a first class citizen I think this is ok. Let me know and I"ll follow up

Reviewed By: philipjameson

shipit-source-id: ee3e5c5c57b030b8ba558fe2ef9617a1c9859770