Fix-up unit tests that hard-coded buck-out path for impending changes to buck-out
Summary:
With the impending changes to buck-out, these unit tests would be broken (Despite everything working as expected). They had an assumed path to buck-out which is obviously not the best way to do this.

This diff changes the way we calculate buck-out paths to call into the code that gets these paths and using that as a reference point to test again.

Reviewed By: stepancheg

shipit-source-id: bef431af1cbe5080c453be3cc37e2b695b302690