Add FormatArg to allow format strings
Summary:
Add a wrapper arg that lets us apply the final stringified value to a format string that has a single %s.

This will be used for an additional post-stringification formatting step. An example is the '/out' parameter to csc. It takes `/out:<some path>`. These cannot be passed as two arguments. However, we don't know the real path until we finally execute the action. Until then it's a SourcePath, or an Artifact, or any number of other things. But in the end, we'd want `/out:buck-out/gen/foo/bar/baz.exe` to be used. In this case the format string would be `/out:%s`. This allows that semi-arbitrary formatting.

Reviewed By: bobyangyf

shipit-source-id: 0693706be63a8ec761e6da1ff1dd9f7aeaf0573e