Add infer_run_info param to rule()
Summary:
Add the 'infer_run_info' parameter to rule(). This signals that if no RunInfo provider is returned, buck should attempt to infer one.

If infer_run_info == true, and no RunInfo is returned, one is inferred from DefaultInfo. If DefaultInfo has zero, or greater than one default outputs, RunInfo is not inferred, and an error is produced. This work is done in a future diff

Reviewed By: bobyangyf

shipit-source-id: 9526f5abf1fb0d3574e012a6b418416e59a101e0