Require @AddToRuleKey on Actions
Summary:
Using the annotation is more consistent with other places that we
compute rulekeys. It's confusing when some places add to rulekey for you and
others require the annotation (ex. wrapping a field in an action in some object
requires adding the annotation).

This just makes it consistent. Because we still want to add all fields of
actions to rulekeys, we enforce that all fields have the annotation.

Reviewed By: bobyangyf

shipit-source-id: f8572c2b63df561cb6afd403aee19f8ea8846cea