Add more paths to csc detection
Summary:
Add an extra path to places to search for csc in integration tests. Bit of a bandaid, but it kind of works until something more comprehensive is put in place.

Also update these tests to allow setup to be run with a difference scenario. This functionality will be used in a future diff

Reviewed By: bobyangyf

shipit-source-id: 3a85030b6650867c949274a124cb48b94ac8a2e5