Create PBXObjectFactory class
Summary:
Writing the project file to disk have been a bit of a long process for two reasons:

1) Generating GIDs for each PBX* object takes awhile (this was resolved by storing the counter variables)
2) We cannot parallelize serializing to disk because of the nested nature of the XML along with the mutability of PBX* objects having their Global ID set on first reference during XML generation.

These next few diffs attempt to attack the second item above.

Right now, `getGlobalID()` for a PBXObject is `null` until `writeProject` is called and the XML is generated via serialization into an NSDictionary object. This means that we can't really parallelize this too well because things are happening serially and we can't guarantee that a gid reference to PBXObject X was created or exists before hand.

Ideally, when we create a PBXObject, we would supply the GID and lock it down, but we really can't do that without breaking the old project version, which I'm hesitant to do in case there are any regressions here.

To solve this, I'm creating a `PBXObjectFactory` wrapper type around calling new on PBXObjects; they simply forward the params to the `new` method of the respective object and then set the global ID then and there. This means that every PBXObject will have a generated global ID once created and we can reference this object anywhere and be able to identify it.

This diff creates this factory type. I'm making it an interface with a default implementation that we will use along with the GIDGenerator one in order to maintain compatibility with buck projectv1 so we can change the APIs; you'll see why in the next diff where PBXObjects can create other PBXObjects.

Reviewed By: williamtwilson

shipit-source-id: fc089c5826ee5cc89693611c5858826e8fa14700